[Refactor] PostRepository 구조 Map으로 변경

+++ src/main/java/org/sopt/Main.java
@@ -44,14 +44,14 @@ public class Main {
                 case "3":
                     System.out.println("\n🔍 [게시글 상세 조회]");
                     System.out.print("📌 조회할 게시글 ID를 입력해주세요: ");
-                    int id = Integer.parseInt(scanner.nextLine());
+                    Long id = Long.parseLong(scanner.nextLine());
                     controller.getPostById(id);
                     break;
 
                 case "4":
                     System.out.println("\n✏️ [게시글 수정]");
                     System.out.print("📌 수정할 게시글 ID를 입력해주세요: ");
-                    int updateId = Integer.parseInt(scanner.nextLine());
+                    Long updateId = Long.parseLong(scanner.nextLine());
                     System.out.print("📝 새 제목을 입력해주세요: ");
                     String newTitle = scanner.nextLine();
                     controller.updatePostTitle(updateId, newTitle);
@@ -60,7 +60,7 @@ public class Main {
                 case "5":
                     System.out.println("\n🗑️ [게시글 삭제]");
                     System.out.print("📌 삭제할 게시글 ID를 입력해주세요: ");
-                    int deleteId = Integer.parseInt(scanner.nextLine());
+                    Long deleteId = Long.parseLong(scanner.nextLine());
                     controller.deletePostById(deleteId);
                     break;
